refactor(api): split the admin router by resource and share the delete set
admin.ts was 1057 lines, 54% of the service, and already sectioned by resource at
its feature-gate comments. It is now one file per resource behind a single
consoleAdminMiddleware mount, in the order the routes were declared, since each
resource's feature gate has to stay above its own routes.

Two extractions come with it. codeSample() replaces nine hand-copied
x-codeSamples blocks: the cast is `as object`, so it contributes nothing to the
route type. deleteSet() takes the allowlist and waitlist deletes, which differed
by eight lines of pure identifier substitution, and owns the transaction, the
write-back and the ordering; the caller still writes its own statement so the
table keeps its Drizzle types. The part worth testing is answerDeleted, which is
pure, lives in batch.ts beside the rest of the set vocabulary, and now has tests.

Verified the RPC surface: api/hono/dist/index.d.mts changes shape, because
.route() composes a union where the chain was an intersection, but every path
keeps its input and output types. web/next type-checks against it, and an
injected bad sort id still fails through InferRequestType, so the client types
are live rather than degraded to any.

No behaviour change. The list endpoints are deliberately left alone: their count
strategies and null ordering differ because their queries differ, not by drift.
